fixed backslash $ expand, need to clean the code
