+++ Makefile
@@ -15,6 +15,7 @@ NAME = minishell
 SRCS = ./main.c \
 	./parsing/check_quotes.c \
 	./parsing/parse.c \
+	./expand/expand.c \
 	./utils/ft_strlen.c \
 	./utils/ft_strcmp.c \
 	./utils/ft_strdup.c \
